@@ -29,39 +29,39 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1",
-    "@angular/cdk": "^6.1",
-    "@angular/common": "^6.1",
-    "@angular/compiler": "^6.1",
-    "@angular/core": "^6.1",
-    "@angular/forms": "^6.1",
-    "@angular/material": "^6.1",
-    "@angular/platform-browser": "^6.1",
-    "@angular/platform-browser-dynamic": "^6.1",
-    "core-js": "^2.4.1",
-    "rxjs": "^6.1.0",
-    "zone.js": "^0.8.26"
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "core-js": "^2.5.7",
+    "rxjs": "^6.6.6",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "^6.1",
-    "@angular/compiler-cli": "^6.1",
-    "@angular/language-service": "^6.1",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "@angular-devkit/build-angular": "~0.803.29",
+    "@angular/cli": "^8.3.29",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~6.14.13",
+    "codelyzer": "^4.5.0",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~4.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^2.4.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "2.7.2",
-    "@angular-devkit/build-angular": "~0.6.0"
+    "ng-packagr": "^10.1.1",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "3.5.3"
   }
 }
