@@ -18,7 +18,7 @@ import { ConfigStoreService } from '../services/config-store.service';
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class CalendarWrapperComponent implements OnChanges {
-  @ViewChild(MatCalendar)
+  @ViewChild(MatCalendar, { static: false })
   matCalendar: MatCalendar<Date>;
 
   @Output()
